move OHazelcastClusterMetadataManager out of OHazelcastPlugin
